Convert recruitment banner unit tests
- move test fixtures.
