+++ spec/fixtures/recruitment_banners.yml
